Only backup .profile if it is a file

+++ install.sh
@@ -4,7 +4,7 @@
 BASEDIR=$(readlink -m `dirname $0`)
 
 # Get rid of existing profile as we want our own.
-mv ~/.profile ~/.profile.bak
+test -f ~/.profile && test -L ~/.profile || mv ~/.profile ~/.profile.bak
 
 for file in profile muttrc vimrc zshrc Xresources gitconfig; do
   test -f ~/.$file || ln -s $BASEDIR/$file ~/.$file
